+++ packages/ledgerstate/utxotest/siglock_test.go
@@ -1,11 +1,11 @@
 package utxotest
 
 import (
-	"github.com/iotaledger/goshimmer/packages/ledgerstate/utxotest/utxodb"
+	"github.com/iotaledger/goshimmer/packages/ledgerstate/utxodb"
 	"testing"
 
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
-	"github.com/iotaledger/goshimmer/packages/ledgerstate/utxotest/utxoutil"
+	"github.com/iotaledger/goshimmer/packages/ledgerstate/utxoutil"
 	"github.com/stretchr/testify/require"
 )
